REF: move find_stack_level to util._exceptions
