+++ pandas/util/_exceptions.py
@@ -1,4 +1,5 @@
 import contextlib
+import inspect
 from typing import Tuple
 
 
@@ -17,3 +18,27 @@ def rewrite_exception(old_name: str, new_name: str):
             args = args + err.args[1:]
         err.args = args
         raise
+
+
+def find_stack_level() -> int:
+    """
+    Find the appropriate stacklevel with which to issue a warning for astype.
+    """
+    stack = inspect.stack()
+
+    # find the lowest-level "astype" call that got us here
+    for n in range(2, 6):
+        if stack[n].function == "astype":
+            break
+
+    while stack[n].function in ["astype", "apply", "_astype"]:
+        # e.g.
+        #  bump up Block.astype -> BlockManager.astype -> NDFrame.astype
+        #  bump up Datetime.Array.astype -> DatetimeIndex.astype
+        n += 1
+
+    if stack[n].function == "__init__":
+        # Series.__init__
+        n += 1
+
+    return n
